test(cypress): fix e2e test jobexecute jobExecute.cy.js
jobExecute独自コマンドをcommands-jobExecute.jsに切り出し
